requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.34
pydantic-settings>=2.7
python-multipart>=0.0.20
pypdf>=5.2
pillow>=11.0
pytesseract>=0.3.13
requests>=2.32

# test
pytest>=9.0
httpx>=0.28
